+++ src/js/sponsors/logo-carousel.js
@@ -28,11 +28,12 @@ function LogoCarousel(containerElement) {
 				return client.getDocument(sponsor.logo.asset._ref)
 			})
 		).then(sponsorLogos => {
+			console.log(sponsorLogos)
 			sponsorsData.map((sponsor, i) => {
 				sponsors.push({
 					name: sponsor.name,
 					websiteUrl: sponsor.website,
-					logoSrc: sponsorLogos[i].url
+					logoSrc: sponsorLogos[i].url + '?w=300'
 				})
 			})
 			sponsors = this.shuffle(sponsors)