+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.6.0] - 2019-10-13
+### Added
+- ACF support for photo albums.
+- Tracy debugger for development.
+
+### Changed
+- Refactored some old code.
+
 ## [1.5.0] - 2019-07-18
 ### Changed
 - Logo carousel slower and no grouping.